requirements.txt
fastapi==0.101.0
fastapi-restful==0.5.0
python-dotenv==1.0.0
uvicorn==0.23.2
torch --index-url https://download.pytorch.org/whl/cu118
torchvision --index-url https://download.pytorch.org/whl/cu118
torchaudio --index-url https://download.pytorch.org/whl/cu118
transformers==4.31.0
diffusers==0.19.3
accelerate==0.21.0
safetensors==0.3.2
